+++ module/CamCap.py
@@ -1,4 +1,5 @@
 import cv2
+import time
 
 """
  キャプチャ開始(Close忘れ注意)
@@ -11,9 +12,19 @@ def OpenCap(rtsp_addr):
     return cap
 
 """
- キャプチャ終了
+ キャプチャ終了(Close後に再度実行しても問題なし)
 """
 def CloseCap(cap):
     # 開いているキャプチャを閉じる
     cap.release()
     print("*** Close Capture! ***")
+
+"""
+  キャプチャを開きなおす
+"""
+def ReOpenCap(cap, rtsp_addr):
+    CloseCap(cap)
+    time.sleep(5)  # 特に意味はないが、すぐにやってもエラーになる気がするので5秒だけ待つ
+    cap = OpenCap(rtsp_addr)
+
+    return cap
