+++ module/gloval_value.py
@@ -1,0 +1,6 @@
+"""
+  グローバル変数を宣言する際に使用するmodule
+  import global_value as g
+  g.val1 = 1
+  のように使用する
+"""
